Backup-services/requirements.txt
google-cloud-storage
python-dotenv
Flask
